requirements.txt
edge-tts
google-api-python-client
google-auth-oauthlib
moviepy
openai
Pillow
pydub
pytest
python-dotenv
PyYAML
